Apply review fixes: trim comments, widen the priced-default guard
